+++ destructify/parsing/__init__.py
@@ -1,3 +1,3 @@
 from .context import *
-from .substream import *
+from .streams import *
 from .expression import this, len_